pyproject.toml
[tool.poetry]
name = "temp_track"
version = "0.1.0"
description = ""
authors = ["Geet Sethi <geetsethi05@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
uagents = "^0.6.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
rich = "^13.6.0"
rich-argparse = "^1.3.0"
fastapi = "^0.103.2"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pylint = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
